[#215] Trim down size of libyottadb.so more than it had bloated in prior #215 commit
As part of a prior #215 commit, the DEFERRED_SIGNAL_HANDLING_CHECK macro was enhanced
to check for Ctrl-Z deferral. Turns out that change caused libyottadb.so size to bloat
from 5.00Mb (in r1.20) to 5.25Mb. This is because this macro is used in heavyweight
macros like REVERT which meant all the code inside this macro gets duplicated across
various parts of the compiled .o files thereby bloating the size of libyottadb.so (which
is just an aggregate of all the individual .o files).

To fix this size bloat, the macro was made minimalistic so it does a simple if check of
a global variable and if that is non-zero, in turn invoke a function that does most of
the code that was previously included in this macro. Thereby we centralize the code in
one place instead of inlining it everywhere the macro is used.

But we also do not want to slow down every macro invocation in the common case (which
is that we do not have any signals deferred). This meant that we cannot fold all the
logic in the macro into the function (as a function call is expensive and we don't want
to do it in every invocation of this macro).

The macro did 3 checks of deferred events
	a) forced_exit
	b) timers
	c) ctrl-z

As part of this commit, all the different checks now update a new global variable
"deferred_signal_handling_needed". It is a bitmask that has one bit set for each type
of deferred event (a, b or c above). So the macro can just do a check of this global
variable and if it is non-zero, invoke a new function "handle_deferred_signal()".

To achieve that, the following was done to the above 3 deferred event scenarios.

a) forced_exit was a multi-state variable so it was not easily replaced by the new bit.
   Therefore "deferred_signal_handling_needed & DEFERRED_SIGNAL_HANDLING_EXIT" was
   set to 1 if ever forced_exit was set to 1 and cleared if forced_exit was 0 or 2.
   And the two were maintained in sync.

b) deferred_timers_check_needed was a boolean variable so it was easily replaced by
   the new bit "deferred_signal_handling_needed & DEFERRED_SIGNAL_HANDLING_TIMERS".

c) suspend_status was a tri-state variable so it was not easily replaced by
   the new bit. Instead it was split into two new variables. One is the new bit
   "deferred_signal_handling_needed & DEFERRED_SIGNAL_HANDLING_CTRLZ" which indicates
   whether a ctrlz event got deferred or not. Another is a new global "is_suspended"
   which is TRUE if the process is currently suspended. This took case of the NOW_SUSPEND
   scenario from before.

With the above changes, the size of libyottadb.so came down from 5.25Mb down to 4.68Mb
which is better than even the 5.00Mb size in r1.20.

While at this, a use of the DEFERRED_SIGNAL_HANDLING_CHECK macro in t_end.c and tp_tend.c
is now removed since it occurs right after a REVERT which has the same check anyways
(REVERT -> ENABLE_INTERRUPTS -> DEFERRED_SIGNAL_HANDLING_CHECK_TRIMMED).