Fix DBFREEZEON/DBFREEZEOFF messages to print full path of database file name (instead of region name)
With multiple database instances running in the same system, the region name is not
necessarily unique across instances and having it in the syslog message does not help
track the message back to a specific database file. Hence it is replaced with the
full path of the database file name corresponding to that region.