[V63005] Enhance %CONVBASEUTIL so it converts as much of input number as it can instead of returning 0
This ensures mpt_extra/conv subtest passes for cases added as part of YottaDB/YottaDB#44.
As part of GTM-5574 changes in GT.M V6.3-005, the conversion function was changed to return 0
in case only a substring of the input number was valid. Whereas YottaDB had previously enhanced
all the numeric conversion routines to convert as much of the input number as possible. Continuing
in that trend, the new %CONVBASEUTIL M program is now enhanced to convert as much as it can.