[V63005] Fixes needed to get scantypedefs script working
The scantypedefs script that we have access to expects every character array member in a structure
to have a size of at least 1 byte. In GT.M V6.3-005, a char array size was changed from 1 to 0 bytes
and so this is reverted back to 1 byte to let the scantypedefs script work fine without any more changes.