[V63005] Fix v63003/gtm8787 subtest failure due to a regression in V6.3-005 where NOPRINCIO error no longer showed up in syslog when MUPIP LOAD process lost its terminal/principal-device
The primary issue was in GTM_FWRITE macro where rc was being assigned to the return value of ferror().
I think this was done with the intent that ferror() will return the "errno" of the fwrite() call.
But ferror() returns a non-zero value in case of an error and the actual value it returned was 1
in case of an ENO5/EIO. So rc had the value of 1 which to the caller gtm_fprintf() incorrectly signaled a
successful call to fprintf() resulting in the NOPRINCIO error not being written to syslog.