[V63005] Fix off-by-one buffer overrun (missed out by GTM-9001 in V6.3-005)
The unicode/dse subtest (with dbg) failed as follows only on the arm boxes (not on x86_64).

DSE> %YDB-F-ASSERT, Assert failed in sr_port/gtm_malloc_src.h line 904
	 for expression (0 == memcmp(trailerMarker, markerChar, SIZEOF(markerChar)))

It is an off-by-one buffer overflow in DSE and is a portable issue. Fixes in this area were
done in GT.M V6.3-005 as part of GTM-9001 but look like they were incomplete.

The off-by-one buffer overrun is now fixed in 2 more functions (cli_str_setup and cli_lex_setup).
GTM-9001 did it only in 1 function (cli_lex_in_expand).