[#318] Changes to ZSYSTEM_FILTER/PIPE_FILTER in restrict.txt are ignored if they happen within 1 second
M labelrefs specified in ZSYSTEM_FILTER/PIPE_FILTER in $ydb_dist/restrict.txt were stored in a table format
suitable for a call-in invocation in the file $ydb_dist/filter_commands.tab. Any newer changes to restrict.txt
relative to filter_commands.tab required recreating filter_commands.tab. This check was done using 1-second
level timestamp granularity which is too coarse. This is now fixed to do nanosecond-level timestamp checking.
That ensures multiple changes to restrict.txt within the same second result in recreation of filter_commands.tab.

In addition, if the timestamp matches at the nanosecond level, a recreation is done (whereas previously
if the timestamp at the second level matched, no recreation was done). This is a just-to-be-safe change.