[#321] Journal records fed to replication filters (external filters) include timestamps
External filters (invoked from either the source or receiver server) are provided the
journal record in the journal extract format. Until now, the timestamps in the journal
extract format was zeroed out before passing it to the external filter program. But some
filters might find the need for the timestamp to decide how to act on the journal record
so pass the non-zero timestamp (in $HOROLOG format) to the external filter unconditionally.