[#324] Error inside indirection usage in direct mode using $ETRAP (not $ZTRAP) causes process to incorrectly terminate (#325)
The way direct mode check was done in mdb_condition_handler() was to check if the frame
immediately previous to the current frame is a direct-mode frame (SFT_DM) or if we are in compile
time and if so check if the current frame is a direct-mode frame. This meant that if an error
occurs inside an indirection usage (which translates to an arbitrary number of uncounted frames
underneath the SFT_DM frame), we would end up treating that error as coming from a non-direct-mode
action which is inconsistent since the same error outside of the indirection will be treated as if
it happened in direct-mode. This inconsistency is fixed by checking all frames back from the current
one until we see either a counted frame or a SFF_INDCE frame and SFT_DM frame sequence. And using that
to decide whether we are in direct mode or not.

As part of investigating a failure in a prior state of the code fixes, noticed a seemingly incorrect
&& (logical AND) usage in sr_port/trans_code_cleanup.c and fixed it to & (bitwise AND) since
SFF_INDCE is a bitmask.