[#335] fix CMakeLists.txt to allow building with Ninja (#336)
The cmake copy command does not support wildcards (https://cmake.org/cmake/help/v3.12/manual/cmake.1.html, and https://gitlab.kitware.com/cmake/cmake/merge_requests/687/diffs); in some cases, wildcards may work because the generator calls a shell which expands the wildcard, but this is not always true. In the case of Ninja, the wildcard is evidently not expanded by the shell.

Because Ninja orders things differently than the Unix Makefile generator did, we got unlucky in that the help files happen to get created before the M files are copied from the source directories. Adding a dependency to the help files to ensure the M files are placed before the help routines are generated fixed this issue.