[GTM-5574] Fix %OD and %OH to not issue incorrect NUMOFLOW error in case a huge octal number (>= 10**47) is specified (#337)
This was mentioned as having been fixed in GTM-5574 in GT.M V6.3-005 but we noticed during
testing that it was not. Only %OD and %OH seemed to have the NUMOFLOW issue. The others
(e.g. %HO, %HD etc.) worked fine.

The check for whether a number is negative can result in NUMOFLOW error if done as a numeric
expression (i.e. if (x<0)). Instead we now check if the first digit is a '-'. That has no numeric aspect
to it and therefore works without a NUMOFLOW error while at the same time doing the negative
number check.